requirements.txt
alembic==1.18.4
annotated-types==0.7.0
attrs==23.2.0
backports.zstd @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_backports.zstd_1767044984/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1764016952863/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1772001073725/work/certifi
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1772816396005/work
contourpy==1.3.3
cycler==0.12.1
fastrlock==0.8.3
fonttools==4.62.0
greenlet==3.3.2
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
implicit==0.7.2
joblib==1.5.3
kiwisolver==1.5.0
lightfm @ file:///home/conda/feedstock_root/build_artifacts/lightfm_1679649519722/work
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1707225376651/work/dist/numpy-1.26.4-cp311-cp311-linux_x86_64.whl#sha256=d08e1c9e5833ae7780563812aa73e2497db1ee3bd5510d3becb8aa18aa2d0c7c
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==2.3.3
pillow==12.1.1
pm-implicit==0.7.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
RecTools==0.18.0
requests @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_requests_1766926974/work
scikit-learn==1.8.0
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1771879143720/work/dist/scipy-1.17.1-cp311-cp311-linux_x86_64.whl#sha256=f015f9f22d3862e43fde9957ff53c8c012eaed642484f10b48295cd9d85ab732
six==1.17.0
SQLAlchemy==2.0.48
threadpoolctl==3.6.0
tqdm==4.67.3
typeguard==4.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1767817748113/work
fastapi==0.115.11
uvicorn==0.34.0
